adding a base activity to set the action bar consistently across the different screens
