+++ app/src/main/java/com/greenaddress/abcore/ABCoreBaseActivity.java
@@ -1,0 +1,56 @@
+package com.greenaddress.abcore;
+
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.preference.PreferenceManager;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+public class ABCoreBaseActivity extends AppCompatActivity {
+
+    @Override
+    public boolean onCreateOptionsMenu(final Menu menu) {
+        final MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.menu, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(final MenuItem item) {
+        // Handle item selection
+        switch (item.getItemId()) {
+            case R.id.main:
+                startActivity(new Intent(this, MainActivity.class));
+                return true;
+            case R.id.configuration:
+                startActivity(new Intent(this, SettingsActivity.class));
+                return true;
+            case R.id.peerview:
+                startActivity(new Intent(this, PeerActivity.class));
+                return true;
+            case R.id.debug:
+                startActivity(new Intent(this, LogActivity.class));
+                return true;
+            case R.id.console:
+                startActivity(new Intent(this, ConsoleActivity.class));
+                return true;
+            case R.id.about:
+                startActivity(new Intent(this, AboutActivity.class));
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    protected void setUpTopNavBar(){
+        final Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
+        final String useDistribution = prefs.getString("usedistribution", "core");
+        getSupportActionBar().setSubtitle(getString(R.string.subtitle, useDistribution));
+    }
+}
