+++ app/src/main/res/layout/activity_peer.xml
@@ -22,7 +22,42 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <include layout="@layout/content_peer"/>
+    <RelativeLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingBottom="@dimen/activity_vertical_margin"
+        android:paddingLeft="@dimen/activity_horizontal_margin"
+        android:paddingRight="@dimen/activity_horizontal_margin"
+        android:paddingTop="@dimen/activity_vertical_margin"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior">
+
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:textAppearance="?android:attr/textAppearanceMedium"
+            android:text="IP:Port - /Version/ - Block Height"
+            android:id="@+id/textView2" android:layout_alignParentTop="true"
+            android:layout_alignParentLeft="true" android:layout_alignParentStart="true"
+            android:layout_alignRight="@android:id/list" android:layout_alignEnd="@android:id/list"/>
+
+        <ListView android:id="@+id/listview"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:paddingTop="@dimen/activity_vertical_margin"
+            android:drawSelectorOnTop="false" android:layout_below="@+id/textView2"/>
+
+
+        <ProgressBar
+            style="?android:attr/progressBarStyleLarge"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/progressBarPeerList" android:layout_centerInParent="true"
+            android:layout_centerVertical="true"  android:layout_centerHorizontal="true"/>
+
+
+    </RelativeLayout>
 
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/fab"
