+++ app/src/main/res/menu/menu.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/main"
+          android:title="Dashboard"/>
     <item android:id="@+id/configuration"
           android:title="Settings"/>
     <item android:id="@+id/peerview"
